Rename QueryService to Statistics::QueryService
This better reflects it's scope as specifically for statistics.
We may want to repurpose a general Hyrax::QueryService for the Valkyrie
integration.
